telegram-bot/requirements.txt
python-telegram-bot==20.7
openai==1.3.0
python-dotenv==1.0.0
atlassian-python-api==3.41.0  # Для Jira